fix(synch): fix a bug in `sema_up`
The waiters list might be not ordered if the semaphore is not managed by
locking mechanisms. Of course, the best way of solving this issue is
using data structures such as rbtrees, but quick and dirty hack will be
used for now.
Maybe fast-path for lock-managed semaphores can be added with small
effort.